+++ DOCKER/Makefile
@@ -13,10 +13,4 @@ build_testing:
 push_develop:
 	docker push "tendermint/tendermint:develop"
 
-build_remote_val_harness:
-	@IMAGE=remote_val_harness DOCKERFILE=Dockerfile.remote_val_harness sh -c "'$(CURDIR)/build.sh'"
-
-push_remote_val_harness:
-	@IMAGE=remote_val_harness sh -c "'$(CURDIR)/push.sh'"
-
 .PHONY: build build_develop push push_develop
