+++ DOCKER/build.sh
@@ -1,10 +1,6 @@
 #!/usr/bin/env bash
 set -e
 
-# By default, we build Tendermint's Docker image
-IMAGE=${IMAGE:-"tendermint"}
-DOCKERFILE=${DOCKERFILE:-"Dockerfile"}
-
 # Get the tag from the version, or try to figure it out.
 if [ -z "$TAG" ]; then
 	TAG=$(awk -F\" '/TMCoreSemVer =/ { print $2; exit }' < ../version/version.go)
@@ -20,8 +16,5 @@ read -p "==> Build 3 docker images with the following tags (latest, $TAG, $TAG_N
 echo
 if [[ $REPLY =~ ^[Yy]$ ]]
 then
-		docker build -t "tendermint/${IMAGE}" \
-			-t "tendermint/${IMAGE}:$TAG" \
-			-t "tendermint/${IMAGE}:$TAG_NO_PATCH" \
-			-f ${DOCKERFILE} .
+		docker build -t "tendermint/tendermint" -t "tendermint/tendermint:$TAG" -t "tendermint/tendermint:$TAG_NO_PATCH" .
 fi
