+++ tools/tm-signer-harness/internal/test_harness.go
