+++ tools/tm-signer-harness/internal/test_harness_test.go
