+++ tools/tm-signer-harness/internal/utils.go
